+++ integration/crud-typeorm/notes/notes.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { TypeOrmCrudService } from '../../../packages/crud-typeorm/src';
+
+import { Note } from './note.entity';
+
+@Injectable()
+export class NotesService extends TypeOrmCrudService<Note> {
+  constructor(@InjectRepository(Note) repo) {
+    super(repo);
+  }
+}