+++ packages/crud-typeorm/test/c.basic-crud.spec.ts
@@ -17,6 +17,8 @@ import { CompaniesService } from './__fixture__/companies.service';
 import { UsersService } from './__fixture__/users.service';
 import { DevicesService } from './__fixture__/devices.service';
 
+const isMysql = process.env.TYPEORM_CONNECTION === 'mysql';
+
 // tslint:disable:max-classes-per-file no-shadowed-variable
 describe('#crud-typeorm', () => {
   describe('#basic crud using alwaysPaginate default respects global limit', () => {
@@ -32,15 +34,15 @@ describe('#crud-typeorm', () => {
         limit: 3,
       },
     })
-    @Controller('companies')
-    class CompaniesController {
+    @Controller('companies0')
+    class CompaniesController0 {
       constructor(public service: CompaniesService) {}
     }
 
     beforeAll(async () => {
       const fixture = await Test.createTestingModule({
         imports: [TypeOrmModule.forRoot(withCache), TypeOrmModule.forFeature([Company])],
-        controllers: [CompaniesController],
+        controllers: [CompaniesController0],
         providers: [
           { provide: APP_FILTER, useClass: HttpExceptionFilter },
           CompaniesService,
@@ -65,7 +67,7 @@ describe('#crud-typeorm', () => {
     describe('#getAllBase', () => {
       it('should return an array of all entities', (done) => {
         return request(server)
-          .get('/companies')
+          .get('/companies0')
           .end((_, res) => {
             expect(res.status).toBe(200);
             expect(res.body.data.length).toBe(3);
@@ -236,6 +238,18 @@ describe('#crud-typeorm', () => {
     }
 
     @Crud({
+      model: { type: User },
+      params: {
+        companyId: { field: 'companyId', type: 'number', primary: true },
+        profileId: { field: 'profileId', type: 'number', primary: true },
+      },
+    })
+    @Controller('users4')
+    class UsersController4 {
+      constructor(public service: UsersService) {}
+    }
+
+    @Crud({
       model: { type: Device },
       params: {
         deviceKey: {
@@ -266,6 +280,7 @@ describe('#crud-typeorm', () => {
           UsersController,
           UsersController2,
           UsersController3,
+          UsersController4,
           DevicesController,
         ],
         providers: [
@@ -353,13 +368,22 @@ describe('#crud-typeorm', () => {
           });
       });
       it('should return an entities with offset', (done) => {
-        const query = qb.setOffset(3).query();
+        const queryObj = qb.setOffset(3);
+        if (isMysql) {
+          queryObj.setLimit(10);
+        }
+        const query = queryObj.query();
         return request(server)
           .get('/companies')
           .query(query)
           .end((_, res) => {
             expect(res.status).toBe(200);
-            expect(res.body.length).toBe(7);
+            if (isMysql) {
+              expect(res.body.count).toBe(7);
+              expect(res.body.data.length).toBe(7);
+            } else {
+              expect(res.body.length).toBe(7);
+            }
             done();
           });
       });
@@ -392,6 +416,15 @@ describe('#crud-typeorm', () => {
             expect(res.status).toBe(200);
             expect(res.body.id).toBe(1);
             expect(res.body.domain).toBeTruthy();
+            done();
+          });
+      });
+      it('should return an entity with compound key', (done) => {
+        return request(server)
+          .get('/users4/1/5')
+          .end((_, res) => {
+            expect(res.status).toBe(200);
+            expect(res.body.id).toBe(5);
             done();
           });
       });
@@ -579,7 +612,7 @@ describe('#crud-typeorm', () => {
     describe('#deleteOneBase', () => {
       it('should return status 404', (done) => {
         return request(server)
-          .delete('/companies/333')
+          .delete('/companies/3333')
           .end((_, res) => {
             expect(res.status).toBe(404);
             done();