+++ packages/crud/src/crud/crud-routes.factory.ts
@@ -10,6 +10,7 @@ import {
   isEqual,
   getOwnPropNames,
   isNil,
+  isUndefined,
 } from '@nestjsx/util';
 import * as deepmerge from 'deepmerge';
 
@@ -73,8 +74,9 @@ export class CrudRoutesFactory {
     // merge auth config
     const authOptions = R.getCrudAuthOptions(this.target);
     this.options.auth = isObjectFull(authOptions) ? authOptions : {};
-    this.options.auth.property =
-      this.options.auth.property || CrudConfigService.config.auth.property;
+    if (isUndefined(this.options.auth.property)) {
+      this.options.auth.property = CrudConfigService.config.auth.property;
+    }
 
     // merge query config
     const query = isObjectFull(this.options.query) ? this.options.query : {};
@@ -92,7 +94,7 @@ export class CrudRoutesFactory {
       : isObjectFull(CrudConfigService.config.params)
       ? CrudConfigService.config.params
       : {};
-    const hasPrimary = this.getPrimaryParam();
+    const hasPrimary = this.getPrimaryParams().length > 0;
     if (!hasPrimary) {
       this.options.params['id'] = {
         field: 'id',
@@ -281,7 +283,9 @@ export class CrudRoutesFactory {
   }
 
   private createRoutes(routesSchema: BaseRoute[]) {
-    const primaryParam = this.getPrimaryParam();
+    const primaryParams = this.getPrimaryParams().filter(
+      (param) => !this.options.params[param].disabled,
+    );
 
     routesSchema.forEach((route) => {
       if (this.canCreateRoute(route.name)) {
@@ -292,8 +296,8 @@ export class CrudRoutesFactory {
         this.setBaseRouteMeta(route.name);
       }
 
-      if (route.withParams && !this.options.params[primaryParam].disabled) {
-        route.path = `/:${primaryParam}`;
+      if (route.withParams && primaryParams.length > 0) {
+        route.path = primaryParams.map((param) => `/:${param}`).join('');
       }
     });
   }
@@ -389,8 +393,8 @@ export class CrudRoutesFactory {
     }
   }
 
-  private getPrimaryParam(): string {
-    return objKeys(this.options.params).find(
+  private getPrimaryParams(): string[] {
+    return objKeys(this.options.params).filter(
       (param) => this.options.params[param] && this.options.params[param].primary,
     );
   }
@@ -488,7 +492,7 @@ export class CrudRoutesFactory {
           {},
         )
       : this.options.params;
-    const pathParamsMeta = Swagger.createPathParasmMeta(params);
+    const pathParamsMeta = Swagger.createPathParamsMeta(params);
     Swagger.setParams([...metadata, ...pathParamsMeta], this.targetProto[name]);
   }