+++ packages/crud/src/interfaces/params-options.interface.ts
@@ -1,3 +1,4 @@
+import { SwaggerEnumType } from '@nestjs/swagger/dist/types/swagger-enum.type';
 import { ParamOptionType } from '@nestjsx/crud-request';
 
 export interface ParamsOptions {
@@ -7,6 +8,7 @@ export interface ParamsOptions {
 export interface ParamOption {
   field?: string;
   type?: ParamOptionType;
+  enum?: SwaggerEnumType;
   primary?: boolean;
   disabled?: boolean;
 }