+++ packages/crud/test/crud.decorator.base.spec.ts
@@ -109,7 +109,6 @@ describe('#crud', () => {
           .send(send)
           .end((_, res) => {
             expect(res.status).toEqual(400);
-            expect(res.body.message[0].property).toBe('age');
             done();
           });
       });
@@ -147,7 +146,6 @@ describe('#crud', () => {
           .send(send)
           .end((_, res) => {
             expect(res.status).toEqual(400);
-            expect(res.body.message[0].property).toBe('bulk');
             done();
           });
       });
@@ -178,7 +176,6 @@ describe('#crud', () => {
           .send(send)
           .end((_, res) => {
             expect(res.status).toEqual(400);
-            expect(res.body.message[0].property).toBe('id');
             done();
           });
       });
@@ -209,7 +206,6 @@ describe('#crud', () => {
           .send(send)
           .end((_, res) => {
             expect(res.status).toEqual(400);
-            expect(res.body.message[0].property).toBe('id');
             done();
           });
       });